Use Composition
