+++ src/App.tsx
@@ -1,16 +1,7 @@
-import {MuiThemeProvider, createMuiTheme} from '@material-ui/core';
 import * as React from 'react';
+import AppContext from './AppContext';
 import {Content, Footer, Header} from './components/layout/';
-
-const AppContext = React.createContext({
-  state: {
-    inDarkMode: false,
-  },
-});
-
-interface Context {
-  state: State;
-}
+import Theme from './Theme';
 
 interface Props {}
 
@@ -22,7 +13,7 @@ class App extends React.Component<Props, State> {
   constructor(props: Props) {
     super(props);
     this.state = {
-      inDarkMode: true,
+      inDarkMode: false,
     };
   }
 
@@ -33,24 +24,12 @@ class App extends React.Component<Props, State> {
           state: this.state,
         }}
       >
-        <AppContext.Consumer>
-          {(context: Context) => (
-            <MuiThemeProvider
-              theme={createMuiTheme({
-                palette: {
-                  type: context.state.inDarkMode ? 'dark' : 'light',
-                },
-                typography: {
-                  useNextVariants: true,
-                },
-              })}
-            >
-              <Header />
-              <Content />
-              <Footer />
-            </MuiThemeProvider>
-          )}
-        </AppContext.Consumer>
+        <Theme>
+          <Header />
+          <Content />
+          <Footer />
+        </Theme>
+        )}
       </AppContext.Provider>
     );
   }
